Token price: Get price form derivedETH field
